(FIX) Accomodate the upstream changes to the VM Pooler API
This is sort of a brute-force fix to work around a new quirk in the
VM Pooler API -- the actual platform name is now the key of note,
not the platform name you asked for. If a platform isn't aliased,
they're functionally identical but if the platform is aliased, then
this is a breaking change.